# [libc++] Speed up classic locale