--- a/libcxx/benchmarks/stringstream.bench.cpp
+++ b/libcxx/benchmarks/stringstream.bench.cpp
@@ -6,7 +6,7 @@
 
 TEST_NOINLINE double istream_numbers();
 
-double istream_numbers(std::locale* l) {
+double istream_numbers(std::locale* loc) {
   const char* a[] = {"-6  69 -71  2.4882e-02 -100 101 -2.00005 5000000 -50000000",
                      "-25 71   7 -9.3262e+01 -100 101 -2.00005 5000000 -50000000",
                      "-14 53  46 -6.7026e-02 -100 101 -2.00005 5000000 -50000000"};
@@ -15,8 +15,8 @@ double istream_numbers(std::locale* l) {
   double f1 = 0.0, f2 = 0.0, q = 0.0;
   for (int i = 0; i < 3; i++) {
     std::istringstream s(a[i]);
-    if (l)
-      s.imbue(*l);
+    if (loc)
+      s.imbue(*loc);
     s >> a1 >> a2 >> a3 >> f1 >> a4 >> a5 >> f2 >> a6 >> a7;
     q += (a1 + a2 + a3 + a4 + a5 + a6 + a7 + f1 + f2) / 1000000;
   }
@@ -28,8 +28,8 @@ struct LocaleSelector {
   std::locale old;
 
   LocaleSelector(benchmark::State& state) {
-    static std::mutex mu;
-    std::lock_guard l(mu);
+    static std::mutex mutex;
+    std::lock_guard guard(mutex);
     switch (state.range(0)) {
     case 0: {
       old   = std::locale::global(std::locale::classic());
@@ -38,14 +38,14 @@ struct LocaleSelector {
     }
     case 1: {
       old = std::locale::global(std::locale::classic());
-      thread_local std::locale l("en_US.UTF-8");
-      imbue = &l;
+      thread_local std::locale loc("en_US.UTF-8");
+      imbue = &loc;
       break;
     }
     case 2: {
       old = std::locale::global(std::locale::classic());
-      static std::locale l("en_US.UTF-8");
-      imbue = &l;
+      static std::locale loc("en_US.UTF-8");
+      imbue = &loc;
       break;
     }
     case 3: {
@@ -57,8 +57,8 @@ struct LocaleSelector {
   }
 
   ~LocaleSelector() {
-    static std::mutex mu;
-    std::lock_guard l(mu);
+    static std::mutex mutex;
+    std::lock_guard guard(mutex);
     std::locale::global(old);
   }
 };