@dvyukov 2023-11-13
This is an alternative version of #70631 (without inlining of ctor/dtor).