@ldionne 2023-11-23
I took the freedom of applying the suggestion that doesn't require atomics at all, and rebasing on top of my other refactoring to save you the trouble of conflict resolution. @mordante @dvyukov What do you folks think about this?