@dvyukov 2023-11-23
I agree this is better. The change looks good to me.
+1 for merging